Safari-609.1.11.1/LayoutTests/fast/events/touch/ios/touch-event-regions-expected.txt

touch start touch move touch end touch force change passive
Overflowing
touchstart
250, 50 - 266, 66
50, 100 - 192, 222
250, 400 - 392, 522
touchstart
250, 50 - 266, 66
250, 100 - 392, 222
250, 400 - 392, 522
touchend
250, 50 - 266, 66
50, 250 - 192, 372
250, 400 - 392, 522
touchforcechange
250, 50 - 266, 66
250, 250 - 392, 372
250, 400 - 392, 522
passive
50, 400 - 192, 522
